Touch - Documentation
CDocs for touch() complete

+++ touch.c
@@ -5,7 +5,18 @@
 #include <unistd.h>
 #include <errno.h>
 
-int touch(char *filepath, int opts) {
+/**
+  * method implements all functionality of bash command 'touch'
+  *
+  * @param filepath - a pointer to the path to touch
+  * @param opts - an int, acting as an enum to specify options, where:
+  *   - 0: no options or -am
+  *   - 1: -a
+  *   - 2: -m
+  * 
+  * @returns an int indicating the success or failure of the method
+  */
+int touch(char *filepath, unsigned short opts) {
   int result;
   if ((result = access(filepath, F_OK)) < 0) {
     if (errno == ENOENT) {
